emotion_api/requirements.txt
fastapi
uvicorn
opencv-python-headless
numpy
tensorflow-cpu
python-multipart
